pyproject.toml
[project]
name = "materialshapes"
version = "0.1"
description = "Material shapes in pure python."
authors = [
  { name="Ansh Dadwal", email="anshdadwal298@gmail.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">3.8"
dependencies = [
  "kivy",
  "pycairo",
  "pillow"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
